Don't force `ActionButtonUseKeyDown` to `0`
Some people like to play with action buttons triggering on key down
rather than key up. Tukui was forcing this variable to `0` on reload
making it impossible to play in that mode. The reason was that auras
cannot be cleared with right click. Additionally raid utils markers cannot
be used anymore.

This fix corrects both problems by explicitly registering the buttons
interest in mouse click down and up events, thus making it equally
possible to play with `ActionButtonUseKeyDown` set to 0 or 1.