Add instancing and config to WidgetManager
Adding support for instancing of the WidgetManager. The widgets are held
in a shared table, but each instance only accesses their own part.

Adding support to store configuration tables with the widgets. This
will enable reuse of widget functions for different unit frames.

Add PreCreate and PostCreate hooks for external edits.